+++ AndroidWeek1/app/src/main/java/org/sopt/androidweek/PreferenceActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import org.sopt.androidweek.databinding.ActivityPreferenceBinding
+import org.sopt.androidweek.util.preference.Application
 
 class PreferenceActivity : AppCompatActivity() {
     private lateinit var binding: ActivityPreferenceBinding
@@ -17,8 +18,8 @@ class PreferenceActivity : AppCompatActivity() {
     private fun cancelAutoLogin() {
         binding.btnCancel.setOnClickListener {
             Toast.makeText(this, "자동로그인이 해제되었습니다.", Toast.LENGTH_LONG).show()
-            org.sopt.androidweek.util.Application.preferenceUtil.setString(EMAIL, "")
-            org.sopt.androidweek.util.Application.preferenceUtil.setString(PASSWORD, "")
+            Application.preferenceUtil.setString(EMAIL, "")
+            Application.preferenceUtil.setString(PASSWORD, "")
         }
     }
