+++ AndroidWeek1/app/src/main/java/org/sopt/androidweek/home/HomeFragment.kt
@@ -1,15 +1,13 @@
 package org.sopt.androidweek.home
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.viewModels
 import com.google.android.material.tabs.TabLayoutMediator
 import org.sopt.androidweek.R
 import org.sopt.androidweek.databinding.FragmentHomeBinding
-import org.sopt.androidweek.util.BaseFragment
+import org.sopt.androidweek.util.base.BaseFragment
 
 class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
     private lateinit var homeViewpagerAdapter: HomeViewpagerAdapter
