+++ AndroidWeek1/app/src/main/java/org/sopt/androidweek/home/ProfileFragment.kt
@@ -2,7 +2,6 @@ package org.sopt.androidweek.home
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -10,8 +9,7 @@ import androidx.fragment.app.viewModels
 import org.sopt.androidweek.PreferenceActivity
 import org.sopt.androidweek.R
 import org.sopt.androidweek.databinding.FragmentProfileBinding
-import org.sopt.androidweek.util.BaseFragment
-import org.sopt.androidweek.util.setImage
+import org.sopt.androidweek.util.base.BaseFragment
 
 
 class ProfileFragment : BaseFragment<FragmentProfileBinding>(R.layout.fragment_profile) {
