+++ AndroidWeek1/app/src/main/java/org/sopt/androidweek/util/preference/Application.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidweek.util
+package org.sopt.androidweek.util.preference
 
 import android.app.Application
